Discharge & Knob Rewrite + Small Updates
Wrote Discharge module logic and built it on breadboard using button & LED. Small changes still needed for countdown display and lever instead of knob but main logic is in place.

Wrote Knob module logic and built it on breadboard. Small changes needed to fine to rotary switch and timer but main logic in place and working.

Updated notes in Maze, Password to be accurate and reduce clutter.

Wrote logic in time.h to handle counting down at 1.25 or 1.5 times the speed when 1 or 2 strikes respectively are incurred.