Update Discharge & Strikes Display & WHO Pins
Changed discharge to use a 2 digit 14 segment display connected in series to a max7219 along side the strikes counter. Had to customize the pinning to make it work without a differnt chip. Therefore custom counter was needed for what should display for each letter. (Could be made into an array of bits to shorten code)
 Fixed Who's on First pin assignments