Major Password ReWrite & Comments Cleanup
Removed unneeded comments from buzzer, changed which module is being tested "password = true" and small changes to fritzing files.

COMPLETE REWRITE of the password module. It fully functions using the Serial.print to see the display. An actual 5 digit 14 segment display will be wired up and printed to eventually.

Will also likely need to find a way to reduce the number of buttons from 11 for this 1 module to a smaller number. Will figure out how later :)